requirements.txt
python-telegram-bot[webhooks,callback-data]
openai
python-dotenv
deepl
SQLAlchemy
psycopg2-binary
alembic
Jinja2
dnspython
